Updates to RIME's background processing model
Instead of a constantly-present "background RIME" running in a separate
thread, launch RIME on-demand using concurrent.futures to perform
long-running background processing tasks (currently creating a subset
and anonymising).

Launching RIME as a separate process means that we don't need to rely on
sqlite being compiled with thread-level concurrency support (which isn't
always the case); we are more likely to fit within the operating system
maximum per-process open file limit; and we don't keep a large and
mostly-unused data structure hanging around when it's not required.

It does, however, mean that communications about the required task and
any results must be passed through standard IPC mechanisms, so no direct
modification of Python objects between threads.