Move the RIME server to the RIME package and move the frontend to /rime
Putting the RIME server in the RIME package makes it slightly easier to package, because all the code is in the one place. It also means we can go python -m rime once it's installed.

Moving the frontend to /rime paves the way for the RIME server to also serve a pre-built frontend (i.e. for running in production) from a zipped release (i.e. vite build).